requirements.txt
ultralytics==8.4.14
streamlit==1.54.0
opencv-python-headless==4.10.0.84
numpy==2.4.2
Pillow==12.1.1
reportlab==4.4.10
albumentations==1.3.1
torch==2.5.1
torchvision==0.20.1
pandas==2.3.3
matplotlib==3.10.8
PyYAML==6.0.3